pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "cartero"
version = "0.1.0"
description = "Convert structured YAML summaries into repo-scoped change plans."
requires-python = ">=3.14"
dependencies = ["PyYAML>=6.0,<7"]

[project.scripts]
cartero = "cartero.cli:main"
